Bugs behoben und optimiert

+++ src/de/fh/stud/pacmanVS/Playout.java
@@ -26,14 +26,7 @@ public class Playout extends Thread{
 			phaser.awaitAdvance(phase);	// warte auf anweisung des MCTS threads weiterzuarbeiten
 			phase++;
 			running=true;
-//			try {Thread.currentThread().sleep(100);} catch (InterruptedException e) {} // delay für debugausgaben
-			
-			
-			
-//			System.out.println("recieved: Playout thread id_"+id+" phase="+phaser.getPhase());
-//			System.out.println("Playout simulationstartroud: "+toSimulate.round+"_"+toSimulate.amZug);
-			int score=toSimulate.SimulateGame();
-//			System.out.println("Playout Thread "+id+" phase:"+MCTSphaser.getPhase()+" -> Score: "+score);
+			double score=toSimulate.SimulateGame();
 			switch(id){
 			case 0: tree.WaitScore=score;  		break;
 			case 1: tree.GoWestScore=score;		break;
@@ -42,11 +35,7 @@ public class Playout extends Thread{
 			case 4: tree.GoSouthScore=score;	break;
 			default: System.out.println("invalid Playout ID ="+id);
 			}
-//			System.out.println("PlayoutThread "+id+": habe die Simulation Abgeschlossen");
-			//System.out.println("Playout thread done_"+id);
-			tree.phaser.arrive();	// sage dem MCTS thread das die Berechnung abgeschlossen ist
-//			System.out.println("Playout thread_"+id+" phase= "+phaser.getPhase()+" tree phaser ARRIVED:"+tree.phaser.getArrivedParties()+" UNARRIVED:"+tree.phaser.getUnarrivedParties()+" Phase: "+tree.phaser.getPhase());
-//			System.out.println("mctsphaser registered parties "+MCTSphaser.getRegisteredParties());
+			tree.phaser.arrive();
 		}
